backend/app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==18.2.0
av==13.1.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
cloup==3.0.7
Cython==3.1.0
data==0.4
decorator==5.2.1
deprecation==2.1.0
dotenv==0.9.9
fastapi==0.115.12
frozenlist==1.6.0
funcsigs==1.0.2
future==1.0.0
glcontext==3.0.0
google-auth==2.40.1
google-genai==1.15.0
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
isosurfaces==0.1.2
latex==0.7.0
latext==0.0.7
manim==0.19.0
ManimPango==0.6.0
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
multidict==6.4.3
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pdflatex==0.1.3
pillow==11.2.1
pluggy==1.6.0
postgrest==1.0.1
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
pyglet==2.1.6
pyglm==2.8.2
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
realtime==2.4.3
redis==6.1.0
requests==2.32.3
rich==14.0.0
rsa==4.9.1
scipy==1.15.3
screeninfo==0.8.1
shutilwhich==1.1.0
six==1.17.0
skia-pathops==0.8.0.post2
sniffio==1.3.1
soupsieve==2.7
srt==3.5.3
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
svgelements==1.9.6
tempdir==0.7.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
websockets==14.2
yarl==1.20.0
openai==1.79.0